Add files via upload
This will find next even number from a list of postive numbers . To find odd or even there are comments in the script that highlight how it can be done. if you want to include negative numbers and still find the next postive missing number you could do a small for loop that will either create a new list with positive numbers only  or if you want be lazy you could add and z>=0.